requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
Flask-Migrate==4.0.5
alembic==1.12.1 